Various small bugfixes
- Fix equipment not being unique
- Fix broken glossary parents inventory
- Fix elusive bug with shops null pointering due to opening another inventory
- Added a more robust upgrade chance to sessions
- increased # of kills needed to scale spawning up

+++ src/me/neoblade298/neorogue/equipment/armor/LeatherHelmet.java
@@ -11,7 +11,7 @@ import me.neoblade298.neorogue.session.fight.trigger.Trigger;
 
 public class LeatherHelmet extends Equipment {
 	private static final String ID = "leatherHelmet";
-	private double shields;
+	private int shields;
 	
 	public LeatherHelmet(boolean isUpgraded) {
 		super(ID, "Leather Helmet", isUpgraded, Rarity.COMMON, EquipmentClass.CLASSLESS,
